server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
#openai
requests